+++ models/task.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Task = sequelize.define('Task', {
     text: DataTypes.STRING,
     status: DataTypes.BOOLEAN,
-    order: DataTypes.BIGINT
+    order: DataTypes.DOUBLE
   }, {});
   return Task;
 };
